+++ docs/src/internalComponents/shape.js
@@ -6,8 +6,9 @@ class shape extends DocInfo {
   }
 
   getDescription() {
-    return `Place this within [[&lt;extrudeGeometry&gt;|extrudeGeometry]]
-    or [&lt;resources&gt;](resources).`;
+    return `Place this within [[&lt;extrudeGeometry&gt;|extrudeGeometry]],
+    [[&lt;shapeGeometry&gt;|shapeGeometry]],
+    or [[&lt;resources&gt;|resources]].`;
   }
 
   getAttributesText() {